Initial move to View as derived from BasicView
