+++ cmake/KokkosCore_config.h.in
@@ -51,6 +51,7 @@
 #cmakedefine KOKKOS_ENABLE_COMPLEX_ALIGN
 #cmakedefine KOKKOS_OPT_RANGE_AGGRESSIVE_VECTORIZATION  // deprecated
 #cmakedefine KOKKOS_ENABLE_AGGRESSIVE_VECTORIZATION
+#cmakedefine KOKKOS_ENABLE_IMPL_VIEW_LEGACY
 #cmakedefine KOKKOS_ENABLE_IMPL_MDSPAN
 #cmakedefine KOKKOS_ENABLE_ATOMICS_BYPASS
