+++ containers/unit_tests/TestViewCtorPropEmbeddedDim.hpp
@@ -14,6 +14,7 @@
 //
 //@HEADER
 
+#ifdef KOKKOS_ENABLE_IMPL_VIEW_LEGACY
 #include <cstdio>
 
 #include <gtest/gtest.h>
@@ -183,3 +184,4 @@ TEST(TEST_CATEGORY, viewctorprop_embedded_dim) {
   TestViewCtorProp_EmbeddedDim<TEST_EXECSPACE>::test_vcpt(2, 3);
 }
 }  // namespace Test
+#endif
