+++ core/src/Kokkos_View.hpp
@@ -19,484 +19,28 @@
 static_assert(false,
               "Including non-public Kokkos header files is not allowed.");
 #endif
+#include <Kokkos_Macros.hpp>
+#ifdef KOKKOS_ENABLE_IMPL_VIEW_LEGACY
+#include <View/Kokkos_View_legacy.hpp>
+#else
 #ifndef KOKKOS_VIEW_HPP
 #define KOKKOS_VIEW_HPP
 
-#include <type_traits>
-#include <string>
-#include <algorithm>
-#include <initializer_list>
-
-#include <Kokkos_Core_fwd.hpp>
-#include <Kokkos_HostSpace.hpp>
-#include <Kokkos_MemoryTraits.hpp>
-#include <Kokkos_ExecPolicy.hpp>
-#include <View/Hooks/Kokkos_ViewHooks.hpp>
-
-#include <impl/Kokkos_Tools.hpp>
-#include <impl/Kokkos_Utilities.hpp>
-
-#ifdef KOKKOS_ENABLE_IMPL_MDSPAN
-#include <View/MDSpan/Kokkos_MDSpan_Extents.hpp>
-#include <View/MDSpan/Kokkos_MDSpan_Layout.hpp>
-#include <View/MDSpan/Kokkos_MDSpan_Accessor.hpp>
-#endif
-#include <Kokkos_MinMax.hpp>
-
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
-namespace Kokkos {
-namespace Impl {
-
-template <class DataType>
-struct ViewArrayAnalysis;
-
-template <class DataType, class ArrayLayout,
-          typename ValueType =
-              typename ViewArrayAnalysis<DataType>::non_const_value_type>
-struct ViewDataAnalysis;
-
-template <class, class...>
-class ViewMapping {
- public:
-  enum : bool { is_assignable_data_type = false };
-  enum : bool { is_assignable = false };
-};
-
-template <typename IntType>
-constexpr KOKKOS_INLINE_FUNCTION std::size_t count_valid_integers(
-    const IntType i0, const IntType i1, const IntType i2, const IntType i3,
-    const IntType i4, const IntType i5, const IntType i6, const IntType i7) {
-  static_assert(std::is_integral<IntType>::value,
-                "count_valid_integers() must have integer arguments.");
-
-  return (i0 != KOKKOS_INVALID_INDEX) + (i1 != KOKKOS_INVALID_INDEX) +
-         (i2 != KOKKOS_INVALID_INDEX) + (i3 != KOKKOS_INVALID_INDEX) +
-         (i4 != KOKKOS_INVALID_INDEX) + (i5 != KOKKOS_INVALID_INDEX) +
-         (i6 != KOKKOS_INVALID_INDEX) + (i7 != KOKKOS_INVALID_INDEX);
-}
-
-// FIXME Ideally, we would not instantiate this function for every possible View
-// type. We should be able to only pass "extent" when we use mdspan.
-template <typename View>
-KOKKOS_INLINE_FUNCTION void runtime_check_rank(
-    const View&, const bool is_void_spec, const size_t i0, const size_t i1,
-    const size_t i2, const size_t i3, const size_t i4, const size_t i5,
-    const size_t i6, const size_t i7, const char* label) {
-  (void)(label);
-
-  if (is_void_spec) {
-    const size_t num_passed_args =
-        count_valid_integers(i0, i1, i2, i3, i4, i5, i6, i7);
-    // We either allow to pass as many extents as the dynamic rank is, or
-    // as many extents as the total rank is. In the latter case, the given
-    // extents for the static dimensions must match the
-    // compile-time extents.
-    constexpr int rank            = View::rank();
-    constexpr int dyn_rank        = View::rank_dynamic();
-    const bool n_args_is_dyn_rank = num_passed_args == dyn_rank;
-    const bool n_args_is_rank     = num_passed_args == rank;
-
-    if constexpr (rank != dyn_rank) {
-      if (n_args_is_rank) {
-        size_t new_extents[8] = {i0, i1, i2, i3, i4, i5, i6, i7};
-        for (int i = dyn_rank; i < rank; ++i)
-          if (new_extents[i] != View::static_extent(i)) {
-            KOKKOS_IF_ON_HOST(
-                const std::string message =
-                    "The specified run-time extent for Kokkos::View '" +
-                    std::string(label) +
-                    "' does not match the compile-time extent in dimension " +
-                    std::to_string(i) + ". The given extent is " +
-                    std::to_string(new_extents[i]) + " but should be " +
-                    std::to_string(View::static_extent(i)) + ".\n";
-                Kokkos::abort(message.c_str());)
-            KOKKOS_IF_ON_DEVICE(
-                Kokkos::abort(
-                    "The specified run-time extents for a Kokkos::View "
-                    "do not match the compile-time extents.");)
-          }
-      }
-    }
-
-    if (!n_args_is_dyn_rank && !n_args_is_rank) {
-      KOKKOS_IF_ON_HOST(
-          const std::string message =
-              "Constructor for Kokkos::View '" + std::string(label) +
-              "' has mismatched number of arguments. The number "
-              "of arguments = " +
-              std::to_string(num_passed_args) +
-              " neither matches the dynamic rank = " +
-              std::to_string(dyn_rank) +
-              " nor the total rank = " + std::to_string(rank) + "\n";
-          Kokkos::abort(message.c_str());)
-      KOKKOS_IF_ON_DEVICE(Kokkos::abort("Constructor for Kokkos View has "
-                                        "mismatched number of arguments.");)
-    }
-  }
-}
-
-} /* namespace Impl */
-} /* namespace Kokkos */
+#include <View/Kokkos_BasicView.hpp>
 
 // Class to provide a uniform type
 namespace Kokkos {
 namespace Impl {
 template <class ViewType, int Traits = 0>
 struct ViewUniformType;
-}
-}  // namespace Kokkos
-
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
-namespace Kokkos {
-
-/** \class ViewTraits
- *  \brief Traits class for accessing attributes of a View.
- *
- * This is an implementation detail of View.  It is only of interest
- * to developers implementing a new specialization of View.
- *
- * Template argument options:
- *   - View< DataType >
- *   - View< DataType , Space >
- *   - View< DataType , Space , MemoryTraits >
- *   - View< DataType , ArrayLayout >
- *   - View< DataType , ArrayLayout , Space >
- *   - View< DataType , ArrayLayout , MemoryTraits >
- *   - View< DataType , ArrayLayout , Space , MemoryTraits >
- *   - View< DataType , MemoryTraits >
- */
-
-template <class DataType, class... Properties>
-struct ViewTraits;
-
-template <>
-struct ViewTraits<void> {
-  using execution_space = void;
-  using memory_space    = void;
-  using HostMirrorSpace = void;
-  using array_layout    = void;
-  using memory_traits   = void;
-  using specialize      = void;
-  using hooks_policy    = void;
-};
-
-template <class... Prop>
-struct ViewTraits<void, void, Prop...> {
-  // Ignore an extraneous 'void'
-  using execution_space = typename ViewTraits<void, Prop...>::execution_space;
-  using memory_space    = typename ViewTraits<void, Prop...>::memory_space;
-  using HostMirrorSpace = typename ViewTraits<void, Prop...>::HostMirrorSpace;
-  using array_layout    = typename ViewTraits<void, Prop...>::array_layout;
-  using memory_traits   = typename ViewTraits<void, Prop...>::memory_traits;
-  using specialize      = typename ViewTraits<void, Prop...>::specialize;
-  using hooks_policy    = typename ViewTraits<void, Prop...>::hooks_policy;
-};
-
-template <class HooksPolicy, class... Prop>
-struct ViewTraits<
-    std::enable_if_t<Kokkos::Experimental::is_hooks_policy<HooksPolicy>::value>,
-    HooksPolicy, Prop...> {
-  using execution_space = typename ViewTraits<void, Prop...>::execution_space;
-  using memory_space    = typename ViewTraits<void, Prop...>::memory_space;
-  using HostMirrorSpace = typename ViewTraits<void, Prop...>::HostMirrorSpace;
-  using array_layout    = typename ViewTraits<void, Prop...>::array_layout;
-  using memory_traits   = typename ViewTraits<void, Prop...>::memory_traits;
-  using specialize      = typename ViewTraits<void, Prop...>::specialize;
-  using hooks_policy    = HooksPolicy;
-};
-
-template <class ArrayLayout, class... Prop>
-struct ViewTraits<std::enable_if_t<Kokkos::is_array_layout<ArrayLayout>::value>,
-                  ArrayLayout, Prop...> {
-  // Specify layout, keep subsequent space and memory traits arguments
-
-  using execution_space = typename ViewTraits<void, Prop...>::execution_space;
-  using memory_space    = typename ViewTraits<void, Prop...>::memory_space;
-  using HostMirrorSpace = typename ViewTraits<void, Prop...>::HostMirrorSpace;
-  using array_layout    = ArrayLayout;
-  using memory_traits   = typename ViewTraits<void, Prop...>::memory_traits;
-  using specialize      = typename ViewTraits<void, Prop...>::specialize;
-  using hooks_policy    = typename ViewTraits<void, Prop...>::hooks_policy;
-};
-
-template <class Space, class... Prop>
-struct ViewTraits<std::enable_if_t<Kokkos::is_space<Space>::value>, Space,
-                  Prop...> {
-  // Specify Space, memory traits should be the only subsequent argument.
-
-  static_assert(
-      std::is_same<typename ViewTraits<void, Prop...>::execution_space,
-                   void>::value &&
-          std::is_same<typename ViewTraits<void, Prop...>::memory_space,
-                       void>::value &&
-          std::is_same<typename ViewTraits<void, Prop...>::HostMirrorSpace,
-                       void>::value &&
-          std::is_same<typename ViewTraits<void, Prop...>::array_layout,
-                       void>::value,
-      "Only one View Execution or Memory Space template argument");
-
-  using execution_space = typename Space::execution_space;
-  using memory_space    = typename Space::memory_space;
-  using HostMirrorSpace =
-      typename Kokkos::Impl::HostMirror<Space>::Space::memory_space;
-  using array_layout  = typename execution_space::array_layout;
-  using memory_traits = typename ViewTraits<void, Prop...>::memory_traits;
-  using specialize    = typename ViewTraits<void, Prop...>::specialize;
-  using hooks_policy  = typename ViewTraits<void, Prop...>::hooks_policy;
-};
-
-template <class MemoryTraits, class... Prop>
-struct ViewTraits<
-    std::enable_if_t<Kokkos::is_memory_traits<MemoryTraits>::value>,
-    MemoryTraits, Prop...> {
-  // Specify memory trait, should not be any subsequent arguments
-
-  static_assert(
-      std::is_same<typename ViewTraits<void, Prop...>::execution_space,
-                   void>::value &&
-          std::is_same<typename ViewTraits<void, Prop...>::memory_space,
-                       void>::value &&
-          std::is_same<typename ViewTraits<void, Prop...>::array_layout,
-                       void>::value &&
-          std::is_same<typename ViewTraits<void, Prop...>::memory_traits,
-                       void>::value &&
-          std::is_same<typename ViewTraits<void, Prop...>::hooks_policy,
-                       void>::value,
-      "MemoryTrait is the final optional template argument for a View");
-
-  using execution_space = void;
-  using memory_space    = void;
-  using HostMirrorSpace = void;
-  using array_layout    = void;
-  using memory_traits   = MemoryTraits;
-  using specialize      = void;
-  using hooks_policy    = void;
-};
-
-template <class DataType, class... Properties>
-struct ViewTraits {
- private:
-  // Unpack the properties arguments
-  using prop = ViewTraits<void, Properties...>;
-
-  using ExecutionSpace =
-      std::conditional_t<!std::is_void<typename prop::execution_space>::value,
-                         typename prop::execution_space,
-                         Kokkos::DefaultExecutionSpace>;
-
-  using MemorySpace =
-      std::conditional_t<!std::is_void<typename prop::memory_space>::value,
-                         typename prop::memory_space,
-                         typename ExecutionSpace::memory_space>;
-
-  using ArrayLayout =
-      std::conditional_t<!std::is_void<typename prop::array_layout>::value,
-                         typename prop::array_layout,
-                         typename ExecutionSpace::array_layout>;
-
-  using HostMirrorSpace = std::conditional_t<
-      !std::is_void<typename prop::HostMirrorSpace>::value,
-      typename prop::HostMirrorSpace,
-      typename Kokkos::Impl::HostMirror<ExecutionSpace>::Space>;
-
-  using MemoryTraits =
-      std::conditional_t<!std::is_void<typename prop::memory_traits>::value,
-                         typename prop::memory_traits,
-                         typename Kokkos::MemoryManaged>;
-
-  using HooksPolicy =
-      std::conditional_t<!std::is_void<typename prop::hooks_policy>::value,
-                         typename prop::hooks_policy,
-                         Kokkos::Experimental::DefaultViewHooks>;
-
-  // Analyze data type's properties,
-  // May be specialized based upon the layout and value type
-  using data_analysis = Kokkos::Impl::ViewDataAnalysis<DataType, ArrayLayout>;
-
- public:
-  //------------------------------------
-  // Data type traits:
-
-  using data_type           = typename data_analysis::type;
-  using const_data_type     = typename data_analysis::const_type;
-  using non_const_data_type = typename data_analysis::non_const_type;
-
-  //------------------------------------
-  // Compatible array of trivial type traits:
-
-  using scalar_array_type = typename data_analysis::scalar_array_type;
-  using const_scalar_array_type =
-      typename data_analysis::const_scalar_array_type;
-  using non_const_scalar_array_type =
-      typename data_analysis::non_const_scalar_array_type;
-
-  //------------------------------------
-  // Value type traits:
-
-  using value_type           = typename data_analysis::value_type;
-  using const_value_type     = typename data_analysis::const_value_type;
-  using non_const_value_type = typename data_analysis::non_const_value_type;
-
-  //------------------------------------
-  // Mapping traits:
-
-  using array_layout = ArrayLayout;
-  using dimension    = typename data_analysis::dimension;
-
-  using specialize = std::conditional_t<
-      std::is_void<typename data_analysis::specialize>::value,
-      typename prop::specialize,
-      typename data_analysis::specialize>; /* mapping specialization tag */
-
-  static constexpr unsigned rank         = dimension::rank;
-  static constexpr unsigned rank_dynamic = dimension::rank_dynamic;
-
-  //------------------------------------
-  // Execution space, memory space, memory access traits, and host mirror space.
-
-  using execution_space   = ExecutionSpace;
-  using memory_space      = MemorySpace;
-  using device_type       = Kokkos::Device<ExecutionSpace, MemorySpace>;
-  using memory_traits     = MemoryTraits;
-  using host_mirror_space = HostMirrorSpace;
-  using hooks_policy      = HooksPolicy;
-
-  using size_type = typename MemorySpace::size_type;
-
-  enum { is_hostspace = std::is_same<MemorySpace, HostSpace>::value };
-  enum { is_managed = MemoryTraits::is_unmanaged == 0 };
-  enum { is_random_access = MemoryTraits::is_random_access == 1 };
-
-  //------------------------------------
-};
-
-#ifdef KOKKOS_ENABLE_IMPL_MDSPAN
-namespace Impl {
-struct UnsupportedKokkosArrayLayout;
-
-template <class Traits, class Enabled = void>
-struct MDSpanViewTraits {
-  using mdspan_type = UnsupportedKokkosArrayLayout;
-};
-
-// "Natural" mdspan for a view if the View's ArrayLayout is supported.
-template <class Traits>
-struct MDSpanViewTraits<Traits,
-                        std::void_t<typename Impl::LayoutFromArrayLayout<
-                            typename Traits::array_layout>::type>> {
-  using index_type = std::size_t;
-  using extents_type =
-      typename Impl::ExtentsFromDataType<index_type,
-                                         typename Traits::data_type>::type;
-  using mdspan_layout_type =
-      typename Impl::LayoutFromArrayLayout<typename Traits::array_layout>::type;
-  using accessor_type = Impl::SpaceAwareAccessor<
-      typename Traits::memory_space,
-      Kokkos::default_accessor<typename Traits::value_type>>;
-  using mdspan_type = mdspan<typename Traits::value_type, extents_type,
-                             mdspan_layout_type, accessor_type>;
-};
-}  // namespace Impl
-#endif  // KOKKOS_ENABLE_IMPL_MDSPAN
-
-/** \class View
- *  \brief View to an array of data.
- *
- * A View represents an array of one or more dimensions.
- * For details, please refer to Kokkos' tutorial materials.
- *
- * \section Kokkos_View_TemplateParameters Template parameters
- *
- * This class has both required and optional template parameters.  The
- * \c DataType parameter must always be provided, and must always be
- * first. The parameters \c Arg1Type, \c Arg2Type, and \c Arg3Type are
- * placeholders for different template parameters.  The default value
- * of the fifth template parameter \c Specialize suffices for most use
- * cases.  When explaining the template parameters, we won't refer to
- * \c Arg1Type, \c Arg2Type, and \c Arg3Type; instead, we will refer
- * to the valid categories of template parameters, in whatever order
- * they may occur.
- *
- * Valid ways in which template arguments may be specified:
- *   - View< DataType >
- *   - View< DataType , Layout >
- *   - View< DataType , Layout , Space >
- *   - View< DataType , Layout , Space , MemoryTraits >
- *   - View< DataType , Space >
- *   - View< DataType , Space , MemoryTraits >
- *   - View< DataType , MemoryTraits >
- *
- * \tparam DataType (required) This indicates both the type of each
- *   entry of the array, and the combination of compile-time and
- *   run-time array dimension(s).  For example, <tt>double*</tt>
- *   indicates a one-dimensional array of \c double with run-time
- *   dimension, and <tt>int*[3]</tt> a two-dimensional array of \c int
- *   with run-time first dimension and compile-time second dimension
- *   (of 3).  In general, the run-time dimensions (if any) must go
- *   first, followed by zero or more compile-time dimensions.  For
- *   more examples, please refer to the tutorial materials.
- *
- * \tparam Space (required) The memory space.
- *
- * \tparam Layout (optional) The array's layout in memory.  For
- *   example, LayoutLeft indicates a column-major (Fortran style)
- *   layout, and LayoutRight a row-major (C style) layout.  If not
- *   specified, this defaults to the preferred layout for the
- *   <tt>Space</tt>.
- *
- * \tparam MemoryTraits (optional) Assertion of the user's intended
- *   access behavior.  For example, RandomAccess indicates read-only
- *   access with limited spatial locality, and Unmanaged lets users
- *   wrap externally allocated memory in a View without automatic
- *   deallocation.
- *
- * \section Kokkos_View_MT MemoryTraits discussion
- *
- * \subsection Kokkos_View_MT_Interp MemoryTraits interpretation depends on
- * Space
- *
- * Some \c MemoryTraits options may have different interpretations for
- * different \c Space types.  For example, with the Cuda device,
- * \c RandomAccess tells Kokkos to fetch the data through the texture
- * cache, whereas the non-GPU devices have no such hardware construct.
- *
- * \subsection Kokkos_View_MT_PrefUse Preferred use of MemoryTraits
- *
- * Users should defer applying the optional \c MemoryTraits parameter
- * until the point at which they actually plan to rely on it in a
- * computational kernel.  This minimizes the number of template
- * parameters exposed in their code, which reduces the cost of
- * compilation.  Users may always assign a View without specified
- * \c MemoryTraits to a compatible View with that specification.
- * For example:
- * \code
- * // Pass in the simplest types of View possible.
- * void
- * doSomething (View<double*, Cuda> out,
- *              View<const double*, Cuda> in)
- * {
- *   // Assign the "generic" View in to a RandomAccess View in_rr.
- *   // Note that RandomAccess View objects must have const data.
- *   View<const double*, Cuda, RandomAccess> in_rr = in;
- *   // ... do something with in_rr and out ...
- * }
- * \endcode
- */
-
-}  // namespace Kokkos
-
-namespace Kokkos {
+} /* namespace Impl */
 
 template <class T1, class T2>
 struct is_always_assignable_impl;
 
 template <class... ViewTDst, class... ViewTSrc>
 struct is_always_assignable_impl<Kokkos::View<ViewTDst...>,
-                                 Kokkos::View<ViewTSrc...>> {
+                                 Kokkos::View<ViewTSrc...> > {
   using mapping_type = Kokkos::Impl::ViewMapping<
       typename Kokkos::View<ViewTDst...>::traits,
       typename Kokkos::View<ViewTSrc...>::traits,
@@ -511,7 +55,7 @@ struct is_always_assignable_impl<Kokkos::View<ViewTDst...>,
 template <class View1, class View2>
 using is_always_assignable = is_always_assignable_impl<
     std::remove_reference_t<View1>,
-    std::remove_const_t<std::remove_reference_t<View2>>>;
+    std::remove_const_t<std::remove_reference_t<View2> > >;
 
 template <class T1, class T2>
 inline constexpr bool is_always_assignable_v =
@@ -527,7 +71,7 @@ constexpr bool is_assignable(const Kokkos::View<ViewTDst...>& dst,
                                 typename DstTraits::specialize>;
 
   return is_always_assignable_v<Kokkos::View<ViewTDst...>,
-                                Kokkos::View<ViewTSrc...>> ||
+                                Kokkos::View<ViewTSrc...> > ||
          (mapping_type::is_assignable &&
           ((DstTraits::dimension::rank_dynamic >= 1) ||
            (dst.static_extent(0) == src.extent(0))) &&
@@ -549,74 +93,6 @@ constexpr bool is_assignable(const Kokkos::View<ViewTDst...>& dst,
 
 } /* namespace Kokkos */
 
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
-#include <impl/Kokkos_ViewMapping.hpp>
-
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
-namespace Kokkos {
-
-// FIXME_OPENMPTARGET - The `declare target` is needed for the Intel GPUs with
-// the OpenMPTarget backend
-#if defined(KOKKOS_ENABLE_OPENMPTARGET) && defined(KOKKOS_COMPILER_INTEL_LLVM)
-#pragma omp declare target
-#endif
-
-inline constexpr Kokkos::ALL_t ALL{};
-
-#if defined(KOKKOS_ENABLE_OPENMPTARGET) && defined(KOKKOS_COMPILER_INTEL_LLVM)
-#pragma omp end declare target
-#endif
-
-inline constexpr Kokkos::Impl::WithoutInitializing_t WithoutInitializing{};
-
-inline constexpr Kokkos::Impl::AllowPadding_t AllowPadding{};
-
-/** \brief  Create View allocation parameter bundle from argument list.
- *
- *  Valid argument list members are:
- *    1) label as a "string" or std::string
- *    2) memory space instance of the View::memory_space type
- *    3) execution space instance compatible with the View::memory_space
- *    4) Kokkos::WithoutInitializing to bypass initialization
- *    4) Kokkos::AllowPadding to allow allocation to pad dimensions for memory
- * alignment
- */
-template <class... Args>
-inline Impl::ViewCtorProp<typename Impl::ViewCtorProp<void, Args>::type...>
-view_alloc(Args const&... args) {
-  using return_type =
-      Impl::ViewCtorProp<typename Impl::ViewCtorProp<void, Args>::type...>;
-
-  static_assert(!return_type::has_pointer,
-                "Cannot give pointer-to-memory for view allocation");
-
-  return return_type(args...);
-}
-
-template <class... Args>
-KOKKOS_INLINE_FUNCTION
-    Impl::ViewCtorProp<typename Impl::ViewCtorProp<void, Args>::type...>
-    view_wrap(Args const&... args) {
-  using return_type =
-      Impl::ViewCtorProp<typename Impl::ViewCtorProp<void, Args>::type...>;
-
-  static_assert(!return_type::has_memory_space &&
-                    !return_type::has_execution_space &&
-                    !return_type::has_label && return_type::has_pointer,
-                "Must only give pointer-to-memory for view wrapping");
-
-  return return_type(args...);
-}
-
-} /* namespace Kokkos */
-
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
 namespace Kokkos {
 
 template <class DataType, class... Properties>
@@ -626,38 +102,35 @@ template <class>
 struct is_view : public std::false_type {};
 
 template <class D, class... P>
-struct is_view<View<D, P...>> : public std::true_type {};
+struct is_view<View<D, P...> > : public std::true_type {};
 
 template <class D, class... P>
-struct is_view<const View<D, P...>> : public std::true_type {};
+struct is_view<const View<D, P...> > : public std::true_type {};
 
 template <class T>
 inline constexpr bool is_view_v = is_view<T>::value;
 
 template <class DataType, class... Properties>
-class View : public ViewTraits<DataType, Properties...> {
+class View : public BasicView<DataType, Properties...> {
  private:
   template <class, class...>
   friend class View;
   template <class, class...>
+  friend class BasicView;
+  template <class, class...>
   friend class Kokkos::Impl::ViewMapping;
-
-  using view_tracker_type = Kokkos::Impl::ViewTracker<View>;
-
- public:
-  using traits = ViewTraits<DataType, Properties...>;
-
- private:
-  using map_type =
-      Kokkos::Impl::ViewMapping<traits, typename traits::specialize>;
   template <typename V>
   friend struct Kokkos::Impl::ViewTracker;
-  using hooks_policy = typename traits::hooks_policy;
 
-  view_tracker_type m_track;
-  map_type m_map;
+  using base_t = BasicView<DataType, Properties...>;
 
  public:
+  using traits            = ViewTraits<DataType, Properties...>;
+  using mdspan_type       = typename base_t::mdspan_type;
+  using view_tracker_type = typename base_t::view_tracker_type;
+  using pointer_type      = typename base_t::pointer_type;
+  using reference_type    = typename base_t::reference_type;
+
   //----------------------------------------
   /** \brief  Compatible view of array of scalar types */
   using array_type =
@@ -709,123 +182,58 @@ class View : public ViewTraits<DataType, Properties...> {
   //----------------------------------------
   // Domain rank and extents
 
-  static constexpr Impl::integral_constant<size_t, traits::dimension::rank>
-      rank = {};
-  static constexpr Impl::integral_constant<size_t,
-                                           traits::dimension::rank_dynamic>
+  static constexpr Impl::integral_constant<size_t, base_t::rank()> rank = {};
+  static constexpr Impl::integral_constant<size_t, base_t::rank_dynamic()>
       rank_dynamic = {};
 #ifdef KOKKOS_ENABLE_DEPRECATED_CODE_4
-  enum {Rank KOKKOS_DEPRECATED_WITH_COMMENT("Use rank instead.") =
-            map_type::Rank};
+  enum {Rank KOKKOS_DEPRECATED_WITH_COMMENT("Use rank instead.") = rank()};
 #endif
-
-  template <typename iType>
-  KOKKOS_INLINE_FUNCTION constexpr std::enable_if_t<
-      std::is_integral<iType>::value, size_t>
-  extent(const iType& r) const noexcept {
-    return m_map.extent(r);
-  }
-
-  static KOKKOS_INLINE_FUNCTION constexpr size_t static_extent(
-      const unsigned r) noexcept {
-    return map_type::static_extent(r);
-  }
-
-  template <typename iType>
-  KOKKOS_INLINE_FUNCTION constexpr std::enable_if_t<
-      std::is_integral<iType>::value, int>
-  extent_int(const iType& r) const noexcept {
-    return static_cast<int>(m_map.extent(r));
-  }
 
   KOKKOS_INLINE_FUNCTION constexpr typename traits::array_layout layout()
       const {
-    return m_map.layout();
+    return base_t::m_map.layout();
   }
 
-  //----------------------------------------
-  /*  Deprecate all 'dimension' functions in favor of
-   *  ISO/C++ vocabulary 'extent'.
-   */
-
-  KOKKOS_INLINE_FUNCTION constexpr size_t size() const {
-    return m_map.dimension_0() * m_map.dimension_1() * m_map.dimension_2() *
-           m_map.dimension_3() * m_map.dimension_4() * m_map.dimension_5() *
-           m_map.dimension_6() * m_map.dimension_7();
-  }
-
-  KOKKOS_INLINE_FUNCTION constexpr size_t stride_0() const {
-    return m_map.stride_0();
-  }
-  KOKKOS_INLINE_FUNCTION constexpr size_t stride_1() const {
-    return m_map.stride_1();
-  }
-  KOKKOS_INLINE_FUNCTION constexpr size_t stride_2() const {
-    return m_map.stride_2();
-  }
-  KOKKOS_INLINE_FUNCTION constexpr size_t stride_3() const {
-    return m_map.stride_3();
-  }
-  KOKKOS_INLINE_FUNCTION constexpr size_t stride_4() const {
-    return m_map.stride_4();
-  }
-  KOKKOS_INLINE_FUNCTION constexpr size_t stride_5() const {
-    return m_map.stride_5();
-  }
-  KOKKOS_INLINE_FUNCTION constexpr size_t stride_6() const {
-    return m_map.stride_6();
-  }
-  KOKKOS_INLINE_FUNCTION constexpr size_t stride_7() const {
-    return m_map.stride_7();
-  }
+  KOKKOS_INLINE_FUNCTION constexpr size_t stride_0() const { return stride(0); }
+  KOKKOS_INLINE_FUNCTION constexpr size_t stride_1() const { return stride(1); }
+  KOKKOS_INLINE_FUNCTION constexpr size_t stride_2() const { return stride(2); }
+  KOKKOS_INLINE_FUNCTION constexpr size_t stride_3() const { return stride(3); }
+  KOKKOS_INLINE_FUNCTION constexpr size_t stride_4() const { return stride(4); }
+  KOKKOS_INLINE_FUNCTION constexpr size_t stride_5() const { return stride(5); }
+  KOKKOS_INLINE_FUNCTION constexpr size_t stride_6() const { return stride(6); }
+  KOKKOS_INLINE_FUNCTION constexpr size_t stride_7() const { return stride(7); }
 
   template <typename iType>
   KOKKOS_INLINE_FUNCTION constexpr std::enable_if_t<
       std::is_integral<iType>::value, size_t>
   stride(iType r) const {
-    return (
-        r == 0
-            ? m_map.stride_0()
-            : (r == 1
-                   ? m_map.stride_1()
-                   : (r == 2
-                          ? m_map.stride_2()
-                          : (r == 3
-                                 ? m_map.stride_3()
-                                 : (r == 4
-                                        ? m_map.stride_4()
-                                        : (r == 5
-                                               ? m_map.stride_5()
-                                               : (r == 6
-                                                      ? m_map.stride_6()
-                                                      : m_map.stride_7())))))));
+    // base class doesn't have constraint
+    return base_t::stride(r);
   }
 
   template <typename iType>
   KOKKOS_INLINE_FUNCTION void stride(iType* const s) const {
-    m_map.stride(s);
+    base_t::m_map.stride(s);
   }
 
   //----------------------------------------
   // Range span is the span which contains all members.
-
-  using reference_type = typename map_type::reference_type;
-  using pointer_type   = typename map_type::pointer_type;
 
   enum {
     reference_type_is_lvalue_reference =
         std::is_lvalue_reference<reference_type>::value
   };
 
-  KOKKOS_INLINE_FUNCTION constexpr size_t span() const { return m_map.span(); }
-  KOKKOS_INLINE_FUNCTION bool span_is_contiguous() const {
-    return m_map.span_is_contiguous();
-  }
+  //  KOKKOS_INLINE_FUNCTION constexpr size_t span() const { return
+  //  required_span_size(); } KOKKOS_INLINE_FUNCTION bool span_is_contiguous()
+  //  const {
+  //    return is_exhaustive();
+  //  }
   KOKKOS_INLINE_FUNCTION constexpr bool is_allocated() const {
-    return m_map.data() != nullptr;
+    return data() != nullptr;
   }
   KOKKOS_INLINE_FUNCTION constexpr pointer_type data() const {
-    return m_map.data();
+    return base_t::m_map.data();
   }
 
   //----------------------------------------
@@ -834,176 +242,34 @@ class View : public ViewTraits<DataType, Properties...> {
   KOKKOS_INLINE_FUNCTION
   const Kokkos::Impl::ViewMapping<traits, typename traits::specialize>&
   impl_map() const {
-    return m_map;
+    return base_t::m_map;
   }
   KOKKOS_INLINE_FUNCTION
   const Kokkos::Impl::SharedAllocationTracker& impl_track() const {
-    return m_track.m_tracker;
+    return base_t::m_track.m_tracker;
   }
   //----------------------------------------
 
  private:
-  static constexpr bool is_layout_left =
-      std::is_same<typename traits::array_layout, Kokkos::LayoutLeft>::value;
 
-  static constexpr bool is_layout_right =
-      std::is_same<typename traits::array_layout, Kokkos::LayoutRight>::value;
-
-  static constexpr bool is_layout_stride =
-      std::is_same<typename traits::array_layout, Kokkos::LayoutStride>::value;
-
-  static constexpr bool is_default_map =
-      std::is_void<typename traits::specialize>::value &&
-      (is_layout_left || is_layout_right || is_layout_stride);
-
-#if defined(KOKKOS_ENABLE_DEBUG_BOUNDS_CHECK)
-
-#define KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(...)                               \
-  Kokkos::Impl::runtime_check_memory_access_violation<                      \
-      typename traits::memory_space>(                                       \
-      "Kokkos::View ERROR: attempt to access inaccessible memory space",    \
-      __VA_ARGS__);                                                         \
-  Kokkos::Impl::view_verify_operator_bounds<typename traits::memory_space>( \
-      __VA_ARGS__);
-
+#ifdef KOKKOS_ENABLE_DEBUG_BOUNDS_CHECK
+  template <typename... Is>
+  static KOKKOS_FUNCTION void check_access_member_function_valid_args(
+      Is... is) {
+    static_assert(sizeof...(Is) <= 8 - rank);
+    static_assert(Kokkos::Impl::are_integral<Is...>::value);
+    if (!((is == static_cast<IS>(0)) && ... && true))
+      Kokkos::abort("Extra arguments to Kokkos::access must be zero");
+  }
 #else
-
-#define KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(...)                            \
-  Kokkos::Impl::runtime_check_memory_access_violation<                   \
-      typename traits::memory_space>(                                    \
-      "Kokkos::View ERROR: attempt to access inaccessible memory space", \
-      __VA_ARGS__);
-
-#endif
-
   template <typename... Is>
   static KOKKOS_FUNCTION void check_access_member_function_valid_args(Is...) {
-    static_assert(rank <= sizeof...(Is));
-    static_assert(sizeof...(Is) <= 8);
+    static_assert(sizeof...(Is) <= 8 - rank);
     static_assert(Kokkos::Impl::are_integral<Is...>::value);
   }
-
-  template <typename... Is>
-  static KOKKOS_FUNCTION void check_operator_parens_valid_args(Is...) {
-    static_assert(rank == sizeof...(Is));
-    static_assert(Kokkos::Impl::are_integral<Is...>::value);
-  }
+#endif
 
  public:
-  //------------------------------
-  // Rank 1 default map operator()
-
-  template <typename I0>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<(Kokkos::Impl::always_true<I0>::value &&  //
-                        (1 == rank) && is_default_map && !is_layout_stride),
-                       reference_type>
-      operator()(I0 i0) const {
-    check_operator_parens_valid_args(i0);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0)
-    return m_map.m_impl_handle[i0];
-  }
-
-  template <typename I0>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<(Kokkos::Impl::always_true<I0>::value &&  //
-                        (1 == rank) && is_default_map && is_layout_stride),
-                       reference_type>
-      operator()(I0 i0) const {
-    check_operator_parens_valid_args(i0);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0)
-    return m_map.m_impl_handle[m_map.m_impl_offset.m_stride.S0 * i0];
-  }
-
-  //------------------------------
-  // Rank 1 operator[]
-
-  template <typename I0>
-  KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
-      ((1 == rank) && Kokkos::Impl::are_integral<I0>::value && !is_default_map),
-      reference_type>
-  operator[](I0 i0) const {
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0)
-    return m_map.reference(i0);
-  }
-
-  template <typename I0>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<((1 == rank) && Kokkos::Impl::are_integral<I0>::value &&
-                        is_default_map && !is_layout_stride),
-                       reference_type>
-      operator[](I0 i0) const {
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0)
-    return m_map.m_impl_handle[i0];
-  }
-
-  template <typename I0>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<((1 == rank) && Kokkos::Impl::are_integral<I0>::value &&
-                        is_default_map && is_layout_stride),
-                       reference_type>
-      operator[](I0 i0) const {
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0)
-    return m_map.m_impl_handle[m_map.m_impl_offset.m_stride.S0 * i0];
-  }
-
-  //------------------------------
-  // Rank 2 default map operator()
-
-  template <typename I0, typename I1>
-  KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
-      (Kokkos::Impl::always_true<I0, I1>::value &&  //
-       (2 == rank) && is_default_map &&
-       (is_layout_left || is_layout_right || is_layout_stride)),
-      reference_type>
-  operator()(I0 i0, I1 i1) const {
-    check_operator_parens_valid_args(i0, i1);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1)
-    if constexpr (is_layout_left) {
-      if constexpr (rank_dynamic == 0)
-        return m_map.m_impl_handle[i0 + m_map.m_impl_offset.m_dim.N0 * i1];
-      else
-        return m_map.m_impl_handle[i0 + m_map.m_impl_offset.m_stride * i1];
-    } else if constexpr (is_layout_right) {
-      if constexpr (rank_dynamic == 0)
-        return m_map.m_impl_handle[i1 + m_map.m_impl_offset.m_dim.N1 * i0];
-      else
-        return m_map.m_impl_handle[i1 + m_map.m_impl_offset.m_stride * i0];
-    } else {
-      static_assert(is_layout_stride);
-      return m_map.m_impl_handle[i0 * m_map.m_impl_offset.m_stride.S0 +
-                                 i1 * m_map.m_impl_offset.m_stride.S1];
-    }
-#if defined KOKKOS_COMPILER_INTEL
-    __builtin_unreachable();
-#endif
-  }
-
-  // Rank 0 -> 8 operator() except for rank-1 and rank-2 with default map which
-  // have "inlined" versions above
-
-  template <typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
-      (Kokkos::Impl::always_true<Is...>::value &&  //
-       (2 != rank) && (1 != rank) && (0 != rank) && is_default_map),
-      reference_type>
-  operator()(Is... indices) const {
-    check_operator_parens_valid_args(indices...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, indices...)
-    return m_map.m_impl_handle[m_map.m_impl_offset(indices...)];
-  }
-
-  template <typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<(Kokkos::Impl::always_true<Is...>::value &&  //
-                        ((0 == rank) || !is_default_map)),
-                       reference_type>
-      operator()(Is... indices) const {
-    check_operator_parens_valid_args(indices...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, indices...)
-    return m_map.reference(indices...);
-  }
-
   //------------------------------
   // Rank 0
 
@@ -1012,112 +278,43 @@ class View : public ViewTraits<DataType, Properties...> {
       (Kokkos::Impl::always_true<Is...>::value && (0 == rank)), reference_type>
   access(Is... extra) const {
     check_access_member_function_valid_args(extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, extra...)
-    return m_map.reference();
+    return base_t::operator()();
   }
 
   //------------------------------
   // Rank 1
 
   template <typename I0, typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<(Kokkos::Impl::always_true<I0, Is...>::value &&
-                        (1 == rank) && !is_default_map),
-                       reference_type>
-      access(I0 i0, Is... extra) const {
-    check_access_member_function_valid_args(i0, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, extra...)
-    return m_map.reference(i0);
-  }
-
-  template <typename I0, typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<(Kokkos::Impl::always_true<I0, Is...>::value &&
-                        (1 == rank) && is_default_map && !is_layout_stride),
-                       reference_type>
-      access(I0 i0, Is... extra) const {
-    check_access_member_function_valid_args(i0, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, extra...)
-    return m_map.m_impl_handle[i0];
-  }
-
-  template <typename I0, typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<(Kokkos::Impl::always_true<I0, Is...>::value &&
-                        (1 == rank) && is_default_map && is_layout_stride),
-                       reference_type>
-      access(I0 i0, Is... extra) const {
-    check_access_member_function_valid_args(i0, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, extra...)
-    return m_map.m_impl_handle[m_map.m_impl_offset.m_stride.S0 * i0];
+  KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
+      (Kokkos::Impl::always_true<I0, Is...>::value && (1 == rank)),
+      reference_type>
+  access(I0 i0, Is... extra) const {
+    check_access_member_function_valid_args(extra...);
+    return base_t::operator()(i0);
   }
 
   //------------------------------
   // Rank 2
 
   template <typename I0, typename I1, typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<(Kokkos::Impl::always_true<I0, I1, Is...>::value &&
-                        (2 == rank) && !is_default_map),
-                       reference_type>
-      access(I0 i0, I1 i1, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, extra...)
-    return m_map.reference(i0, i1);
-  }
-
-  template <typename I0, typename I1, typename... Is>
   KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
-      (Kokkos::Impl::always_true<I0, I1, Is...>::value && (2 == rank) &&
-       is_default_map &&
-       (is_layout_left || is_layout_right || is_layout_stride)),
+      (Kokkos::Impl::always_true<I0, I1, Is...>::value && (2 == rank)),
       reference_type>
   access(I0 i0, I1 i1, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, extra...)
-    if constexpr (is_layout_left) {
-      if constexpr (rank_dynamic == 0)
-        return m_map.m_impl_handle[i0 + m_map.m_impl_offset.m_dim.N0 * i1];
-      else
-        return m_map.m_impl_handle[i0 + m_map.m_impl_offset.m_stride * i1];
-    } else if constexpr (is_layout_right) {
-      if constexpr (rank_dynamic == 0)
-        return m_map.m_impl_handle[i1 + m_map.m_impl_offset.m_dim.N1 * i0];
-      else
-        return m_map.m_impl_handle[i1 + m_map.m_impl_offset.m_stride * i0];
-    } else {
-      static_assert(is_layout_stride);
-      return m_map.m_impl_handle[i0 * m_map.m_impl_offset.m_stride.S0 +
-                                 i1 * m_map.m_impl_offset.m_stride.S1];
-    }
-#if defined KOKKOS_COMPILER_INTEL
-    __builtin_unreachable();
-#endif
+    check_access_member_function_valid_args(extra...);
+    return base_t::operator()(i0, i1);
   }
 
   //------------------------------
   // Rank 3
 
   template <typename I0, typename I1, typename I2, typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<(Kokkos::Impl::always_true<I0, I1, I2, Is...>::value &&
-                        (3 == rank) && is_default_map),
-                       reference_type>
-      access(I0 i0, I1 i1, I2 i2, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, extra...)
-    return m_map.m_impl_handle[m_map.m_impl_offset(i0, i1, i2)];
-  }
-
-  template <typename I0, typename I1, typename I2, typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<(Kokkos::Impl::always_true<I0, I1, I2, Is...>::value &&
-                        (3 == rank) && !is_default_map),
-                       reference_type>
-      access(I0 i0, I1 i1, I2 i2, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, extra...)
-    return m_map.reference(i0, i1, i2);
+  KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
+      (Kokkos::Impl::always_true<I0, I1, I2, Is...>::value && (3 == rank)),
+      reference_type>
+  access(I0 i0, I1 i1, I2 i2, Is... extra) const {
+    check_access_member_function_valid_args(extra...);
+    return base_t::operator()(i0, i1, i2);
   }
 
   //------------------------------
@@ -1125,24 +322,11 @@ class View : public ViewTraits<DataType, Properties...> {
 
   template <typename I0, typename I1, typename I2, typename I3, typename... Is>
   KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
-      (Kokkos::Impl::always_true<I0, I1, I2, I3, Is...>::value && (4 == rank) &&
-       is_default_map),
+      (Kokkos::Impl::always_true<I0, I1, I2, I3, Is...>::value && (4 == rank)),
       reference_type>
   access(I0 i0, I1 i1, I2 i2, I3 i3, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, i3, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, i3, extra...)
-    return m_map.m_impl_handle[m_map.m_impl_offset(i0, i1, i2, i3)];
-  }
-
-  template <typename I0, typename I1, typename I2, typename I3, typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
-      (Kokkos::Impl::always_true<I0, I1, I2, I3, Is...>::value && (4 == rank) &&
-       !is_default_map),
-      reference_type>
-  access(I0 i0, I1 i1, I2 i2, I3 i3, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, i3, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, i3, extra...)
-    return m_map.reference(i0, i1, i2, i3);
+    check_access_member_function_valid_args(extra...);
+    return base_t::operator()(i0, i1, i2, i3);
   }
 
   //------------------------------
@@ -1152,26 +336,11 @@ class View : public ViewTraits<DataType, Properties...> {
             typename... Is>
   KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
       (Kokkos::Impl::always_true<I0, I1, I2, I3, I4, Is...>::value &&
-       (5 == rank) && is_default_map),
+       (5 == rank)),
       reference_type>
   access(I0 i0, I1 i1, I2 i2, I3 i3, I4 i4, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, i3, i4, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, i3, i4,
-                                     extra...)
-    return m_map.m_impl_handle[m_map.m_impl_offset(i0, i1, i2, i3, i4)];
-  }
-
-  template <typename I0, typename I1, typename I2, typename I3, typename I4,
-            typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
-      (Kokkos::Impl::always_true<I0, I1, I2, I3, I4, Is...>::value &&
-       (5 == rank) && !is_default_map),
-      reference_type>
-  access(I0 i0, I1 i1, I2 i2, I3 i3, I4 i4, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, i3, i4, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, i3, i4,
-                                     extra...)
-    return m_map.reference(i0, i1, i2, i3, i4);
+    check_access_member_function_valid_args(extra...);
+    return base_t::operator()(i0, i1, i2, i3, i4);
   }
 
   //------------------------------
@@ -1181,26 +350,11 @@ class View : public ViewTraits<DataType, Properties...> {
             typename I5, typename... Is>
   KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
       (Kokkos::Impl::always_true<I0, I1, I2, I3, I4, I5, Is...>::value &&
-       (6 == rank) && is_default_map),
+       (6 == rank)),
       reference_type>
   access(I0 i0, I1 i1, I2 i2, I3 i3, I4 i4, I5 i5, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, i3, i4, i5, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, i3, i4, i5,
-                                     extra...)
-    return m_map.m_impl_handle[m_map.m_impl_offset(i0, i1, i2, i3, i4, i5)];
-  }
-
-  template <typename I0, typename I1, typename I2, typename I3, typename I4,
-            typename I5, typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
-      (Kokkos::Impl::always_true<I0, I1, I2, I3, I4, I5, Is...>::value &&
-       (6 == rank) && !is_default_map),
-      reference_type>
-  access(I0 i0, I1 i1, I2 i2, I3 i3, I4 i4, I5 i5, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, i3, i4, i5, extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, i3, i4, i5,
-                                     extra...)
-    return m_map.reference(i0, i1, i2, i3, i4, i5);
+    check_access_member_function_valid_args(extra...);
+    return base_t::operator()(i0, i1, i2, i3, i4, i5);
   }
 
   //------------------------------
@@ -1210,28 +364,11 @@ class View : public ViewTraits<DataType, Properties...> {
             typename I5, typename I6, typename... Is>
   KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
       (Kokkos::Impl::always_true<I0, I1, I2, I3, I4, I5, I6, Is...>::value &&
-       (7 == rank) && is_default_map),
+       (7 == rank)),
       reference_type>
   access(I0 i0, I1 i1, I2 i2, I3 i3, I4 i4, I5 i5, I6 i6, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, i3, i4, i5, i6,
-                                            extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, i3, i4, i5, i6,
-                                     extra...)
-    return m_map.m_impl_handle[m_map.m_impl_offset(i0, i1, i2, i3, i4, i5, i6)];
-  }
-
-  template <typename I0, typename I1, typename I2, typename I3, typename I4,
-            typename I5, typename I6, typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION std::enable_if_t<
-      (Kokkos::Impl::always_true<I0, I1, I2, I3, I4, I5, I6, Is...>::value &&
-       (7 == rank) && !is_default_map),
-      reference_type>
-  access(I0 i0, I1 i1, I2 i2, I3 i3, I4 i4, I5 i5, I6 i6, Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, i3, i4, i5, i6,
-                                            extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, i3, i4, i5, i6,
-                                     extra...)
-    return m_map.reference(i0, i1, i2, i3, i4, i5, i6);
+    check_access_member_function_valid_args(extra...);
+    return base_t::operator()(i0, i1, i2, i3, i4, i5, i6);
   }
 
   //------------------------------
@@ -1242,32 +379,12 @@ class View : public ViewTraits<DataType, Properties...> {
   KOKKOS_FORCEINLINE_FUNCTION
       std::enable_if_t<(Kokkos::Impl::always_true<I0, I1, I2, I3, I4, I5, I6,
                                                   I7, Is...>::value &&
-                        (8 == rank) && is_default_map),
+                        (8 == rank)),
                        reference_type>
       access(I0 i0, I1 i1, I2 i2, I3 i3, I4 i4, I5 i5, I6 i6, I7 i7,
              Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, i3, i4, i5, i6, i7,
-                                            extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, i3, i4, i5, i6,
-                                     i7, extra...)
-    return m_map
-        .m_impl_handle[m_map.m_impl_offset(i0, i1, i2, i3, i4, i5, i6, i7)];
-  }
-
-  template <typename I0, typename I1, typename I2, typename I3, typename I4,
-            typename I5, typename I6, typename I7, typename... Is>
-  KOKKOS_FORCEINLINE_FUNCTION
-      std::enable_if_t<(Kokkos::Impl::always_true<I0, I1, I2, I3, I4, I5, I6,
-                                                  I7, Is...>::value &&
-                        (8 == rank) && !is_default_map),
-                       reference_type>
-      access(I0 i0, I1 i1, I2 i2, I3 i3, I4 i4, I5 i5, I6 i6, I7 i7,
-             Is... extra) const {
-    check_access_member_function_valid_args(i0, i1, i2, i3, i4, i5, i6, i7,
-                                            extra...);
-    KOKKOS_IMPL_VIEW_OPERATOR_VERIFY(m_track, m_map, i0, i1, i2, i3, i4, i5, i6,
-                                     i7, extra...)
-    return m_map.reference(i0, i1, i2, i3, i4, i5, i6, i7);
+    check_access_member_function_valid_args(extra...);
+    return base_t::operator()(i0, i1, i2, i3, i4, i5, i6, i7);
   }
 
 #undef KOKKOS_IMPL_VIEW_OPERATOR_VERIFY
@@ -1282,33 +399,20 @@ class View : public ViewTraits<DataType, Properties...> {
   View() = default;
 
   KOKKOS_FUNCTION
-  View(const View& other) : m_track(other.m_track), m_map(other.m_map) {
-    KOKKOS_IF_ON_HOST((hooks_policy::copy_construct(*this, other);))
-  }
+  View(const View& other) : base_t(other) {}
 
   KOKKOS_FUNCTION
-  View(View&& other)
-      : m_track{std::move(other.m_track)}, m_map{std::move(other.m_map)} {
-    KOKKOS_IF_ON_HOST((hooks_policy::move_construct(*this, other);))
-  }
+  View(View&& other) : base_t(other) {}
 
   KOKKOS_FUNCTION
   View& operator=(const View& other) {
-    m_map   = other.m_map;
-    m_track = other.m_track;
-
-    KOKKOS_IF_ON_HOST((hooks_policy::copy_assign(*this, other);))
-
+    base_t::operator=(other);
     return *this;
   }
 
   KOKKOS_FUNCTION
   View& operator=(View&& other) {
-    m_map   = std::move(other.m_map);
-    m_track = std::move(other.m_track);
-
-    KOKKOS_IF_ON_HOST((hooks_policy::move_assign(*this, other);))
-
+    base_t::operator=(other);
     return *this;
   }
 
@@ -1322,14 +426,7 @@ class View : public ViewTraits<DataType, Properties...> {
       std::enable_if_t<Kokkos::Impl::ViewMapping<
           traits, typename View<RT, RP...>::traits,
           typename traits::specialize>::is_assignable_data_type>* = nullptr)
-      : m_track(rhs), m_map() {
-    using SrcTraits = typename View<RT, RP...>::traits;
-    using Mapping   = Kokkos::Impl::ViewMapping<traits, SrcTraits,
-                                              typename traits::specialize>;
-    static_assert(Mapping::is_assignable,
-                  "Incompatible View copy construction");
-    Mapping::assign(m_map, rhs.m_map, rhs.m_track.m_tracker);
-  }
+      : base_t(rhs) {}
 
   template <class RT, class... RP>
   KOKKOS_INLINE_FUNCTION std::enable_if_t<
@@ -1338,12 +435,7 @@ class View : public ViewTraits<DataType, Properties...> {
           typename traits::specialize>::is_assignable_data_type,
       View>&
   operator=(const View<RT, RP...>& rhs) {
-    using SrcTraits = typename View<RT, RP...>::traits;
-    using Mapping   = Kokkos::Impl::ViewMapping<traits, SrcTraits,
-                                              typename traits::specialize>;
-    static_assert(Mapping::is_assignable, "Incompatible View copy assignment");
-    Mapping::assign(m_map, rhs.m_map, rhs.m_track.m_tracker);
-    m_track.assign(rhs);
+    base_t::operator=(rhs);
     return *this;
   }
 
@@ -1354,32 +446,7 @@ class View : public ViewTraits<DataType, Properties...> {
   template <class RT, class... RP, class Arg0, class... Args>
   KOKKOS_INLINE_FUNCTION View(const View<RT, RP...>& src_view, const Arg0 arg0,
                               Args... args)
-      : m_track(src_view), m_map() {
-    using SrcType = View<RT, RP...>;
-
-    using Mapping = Kokkos::Impl::ViewMapping<void, typename SrcType::traits,
-                                              Arg0, Args...>;
-
-    using DstType = typename Mapping::type;
-
-    static_assert(
-        Kokkos::Impl::ViewMapping<traits, typename DstType::traits,
-                                  typename traits::specialize>::is_assignable,
-        "Subview construction requires compatible view and subview arguments");
-
-    Mapping::assign(m_map, src_view.m_map, arg0, args...);
-  }
-
-  //----------------------------------------
-  // Allocation tracking properties
-
-  KOKKOS_INLINE_FUNCTION
-  int use_count() const { return m_track.m_tracker.use_count(); }
-
-  inline const std::string label() const {
-    return m_track.m_tracker
-        .template get_label<typename traits::memory_space>();
-  }
+      : base_t(src_view, arg0, args...) {}
 
  public:
   //----------------------------------------
@@ -1390,64 +457,7 @@ class View : public ViewTraits<DataType, Properties...> {
       const Impl::ViewCtorProp<P...>& arg_prop,
       std::enable_if_t<!Impl::ViewCtorProp<P...>::has_pointer,
                        typename traits::array_layout> const& arg_layout)
-      : m_track(), m_map() {
-    // Copy the input allocation properties with possibly defaulted properties
-    // We need to split it in two to avoid MSVC compiler errors
-    auto prop_copy_tmp =
-        Impl::with_properties_if_unset(arg_prop, std::string{});
-    auto prop_copy = Impl::with_properties_if_unset(
-        prop_copy_tmp, typename traits::device_type::memory_space{},
-        typename traits::device_type::execution_space{});
-    using alloc_prop = decltype(prop_copy);
-
-    static_assert(traits::is_managed,
-                  "View allocation constructor requires managed memory");
-
-    if (alloc_prop::initialize &&
-        !alloc_prop::execution_space::impl_is_initialized()) {
-      // If initializing view data then
-      // the execution space must be initialized.
-      Kokkos::Impl::throw_runtime_exception(
-          "Constructing View and initializing data with uninitialized "
-          "execution space");
-    }
-
-#ifdef KOKKOS_ENABLE_DEBUG_BOUNDS_CHECK
-    if constexpr (std::is_same_v<typename traits::array_layout,
-                                 Kokkos::LayoutLeft> ||
-                  std::is_same_v<typename traits::array_layout,
-                                 Kokkos::LayoutRight> ||
-                  std::is_same_v<typename traits::array_layout,
-                                 Kokkos::LayoutStride>) {
-      size_t i0 = arg_layout.dimension[0];
-      size_t i1 = arg_layout.dimension[1];
-      size_t i2 = arg_layout.dimension[2];
-      size_t i3 = arg_layout.dimension[3];
-      size_t i4 = arg_layout.dimension[4];
-      size_t i5 = arg_layout.dimension[5];
-      size_t i6 = arg_layout.dimension[6];
-      size_t i7 = arg_layout.dimension[7];
-
-      const std::string& alloc_name =
-          Impl::get_property<Impl::LabelTag>(prop_copy);
-      Impl::runtime_check_rank(
-          *this, std::is_same<typename traits::specialize, void>::value, i0, i1,
-          i2, i3, i4, i5, i6, i7, alloc_name.c_str());
-    }
-#endif
-
-    Kokkos::Impl::SharedAllocationRecord<>* record = m_map.allocate_shared(
-        prop_copy, arg_layout, Impl::ViewCtorProp<P...>::has_execution_space);
-
-    // Setup and initialization complete, start tracking
-    m_track.m_tracker.assign_allocated_record_to_uninitialized(record);
-  }
-
-  KOKKOS_INLINE_FUNCTION
-  void assign_data(pointer_type arg_data) {
-    m_track.m_tracker.clear();
-    m_map.assign_data(arg_data);
-  }
+      : base_t(arg_prop, arg_layout) {}
 
   // Wrap memory according to properties and array layout
   template <class... P>
@@ -1455,37 +465,7 @@ class View : public ViewTraits<DataType, Properties...> {
       const Impl::ViewCtorProp<P...>& arg_prop,
       std::enable_if_t<Impl::ViewCtorProp<P...>::has_pointer,
                        typename traits::array_layout> const& arg_layout)
-      : m_track()  // No memory tracking
-        ,
-        m_map(arg_prop, arg_layout) {
-    static_assert(
-        std::is_same<pointer_type,
-                     typename Impl::ViewCtorProp<P...>::pointer_type>::value,
-        "Constructing View to wrap user memory must supply matching pointer "
-        "type");
-
-#ifdef KOKKOS_ENABLE_DEBUG_BOUNDS_CHECK
-    if constexpr (std::is_same_v<typename traits::array_layout,
-                                 Kokkos::LayoutLeft> ||
-                  std::is_same_v<typename traits::array_layout,
-                                 Kokkos::LayoutRight> ||
-                  std::is_same_v<typename traits::array_layout,
-                                 Kokkos::LayoutStride>) {
-      size_t i0 = arg_layout.dimension[0];
-      size_t i1 = arg_layout.dimension[1];
-      size_t i2 = arg_layout.dimension[2];
-      size_t i3 = arg_layout.dimension[3];
-      size_t i4 = arg_layout.dimension[4];
-      size_t i5 = arg_layout.dimension[5];
-      size_t i6 = arg_layout.dimension[6];
-      size_t i7 = arg_layout.dimension[7];
-
-      Impl::runtime_check_rank(
-          *this, std::is_same<typename traits::specialize, void>::value, i0, i1,
-          i2, i3, i4, i5, i6, i7, "UNMANAGED");
-    }
-#endif
-  }
+      : base_t(arg_prop, arg_layout) {}
 
   // Simple dimension-only layout
   template <class... P>
@@ -1564,13 +544,7 @@ class View : public ViewTraits<DataType, Properties...> {
   KOKKOS_INLINE_FUNCTION View(
       const view_tracker_type& track,
       const Kokkos::Impl::ViewMapping<Traits, typename Traits::specialize>& map)
-      : m_track(track), m_map() {
-    using Mapping =
-        Kokkos::Impl::ViewMapping<traits, Traits, typename traits::specialize>;
-    static_assert(Mapping::is_assignable,
-                  "Incompatible View copy construction");
-    Mapping::assign(m_map, map, track.m_tracker);
-  }
+      : base_t(track, map) {}
 
   // Construct View from internal shared allocation tracker object and map
   // This is here for backwards compatibility for classes that derive from
@@ -1579,19 +553,13 @@ class View : public ViewTraits<DataType, Properties...> {
   KOKKOS_INLINE_FUNCTION View(
       const typename view_tracker_type::track_type& track,
       const Kokkos::Impl::ViewMapping<Traits, typename Traits::specialize>& map)
-      : m_track(track), m_map() {
-    using Mapping =
-        Kokkos::Impl::ViewMapping<traits, Traits, typename traits::specialize>;
-    static_assert(Mapping::is_assignable,
-                  "Incompatible View copy construction");
-    Mapping::assign(m_map, map, track);
-  }
+      : base_t(track, map) {}
 
   //----------------------------------------
   // Memory span required to wrap these dimensions.
   static constexpr size_t required_allocation_size(
       typename traits::array_layout const& layout) {
-    return map_type::memory_span(layout);
+    return base_t::map_type::memory_span(layout);
   }
 
   static constexpr size_t required_allocation_size(
@@ -1601,7 +569,7 @@ class View : public ViewTraits<DataType, Properties...> {
     static_assert(traits::array_layout::is_extent_constructible,
                   "Layout is not constructible from extent arguments. Use "
                   "overload taking a layout object instead.");
-    return map_type::memory_span(typename traits::array_layout(
+    return base_t::map_type::memory_span(typename traits::array_layout(
         arg_N0, arg_N1, arg_N2, arg_N3, arg_N4, arg_N5, arg_N6, arg_N7));
   }
 
@@ -1666,15 +634,16 @@ class View : public ViewTraits<DataType, Properties...> {
  public:
   static KOKKOS_INLINE_FUNCTION size_t
   shmem_size(typename traits::array_layout const& arg_layout) {
-    return map_type::memory_span(arg_layout) + scratch_value_alignment;
+    return base_t::map_type::memory_span(arg_layout) + scratch_value_alignment;
   }
 
   explicit KOKKOS_INLINE_FUNCTION View(
       const typename traits::execution_space::scratch_memory_space& arg_space,
       const typename traits::array_layout& arg_layout)
-      : View(Impl::ViewCtorProp<pointer_type>(reinterpret_cast<pointer_type>(
-                 arg_space.get_shmem_aligned(map_type::memory_span(arg_layout),
-                                             scratch_value_alignment))),
+      : View(Impl::ViewCtorProp<pointer_type>(
+                 reinterpret_cast<pointer_type>(arg_space.get_shmem_aligned(
+                     base_t::map_type::memory_span(arg_layout),
+                     scratch_value_alignment))),
              arg_layout) {}
 
   explicit KOKKOS_INLINE_FUNCTION View(
@@ -1687,14 +656,15 @@ class View : public ViewTraits<DataType, Properties...> {
       const size_t arg_N5 = KOKKOS_IMPL_CTOR_DEFAULT_ARG,
       const size_t arg_N6 = KOKKOS_IMPL_CTOR_DEFAULT_ARG,
       const size_t arg_N7 = KOKKOS_IMPL_CTOR_DEFAULT_ARG)
-      : View(Impl::ViewCtorProp<pointer_type>(
-                 reinterpret_cast<pointer_type>(arg_space.get_shmem_aligned(
-                     map_type::memory_span(typename traits::array_layout(
-                         arg_N0, arg_N1, arg_N2, arg_N3, arg_N4, arg_N5, arg_N6,
-                         arg_N7)),
-                     scratch_value_alignment))),
-             typename traits::array_layout(arg_N0, arg_N1, arg_N2, arg_N3,
-                                           arg_N4, arg_N5, arg_N6, arg_N7)) {
+      : View(
+            Impl::ViewCtorProp<pointer_type>(
+                reinterpret_cast<pointer_type>(arg_space.get_shmem_aligned(
+                    base_t::map_type::memory_span(typename traits::array_layout(
+                        arg_N0, arg_N1, arg_N2, arg_N3, arg_N4, arg_N5, arg_N6,
+                        arg_N7)),
+                    scratch_value_alignment))),
+            typename traits::array_layout(arg_N0, arg_N1, arg_N2, arg_N3,
+                                          arg_N4, arg_N5, arg_N6, arg_N7)) {
     static_assert(traits::array_layout::is_extent_constructible,
                   "Layout is not constructible from extent arguments. Use "
                   "overload taking a layout object instead.");
@@ -1702,7 +672,6 @@ class View : public ViewTraits<DataType, Properties...> {
 
   //----------------------------------------
   // MDSpan converting constructors
-#ifdef KOKKOS_ENABLE_IMPL_MDSPAN
   template <typename U = typename Impl::MDSpanViewTraits<traits>::mdspan_type>
   KOKKOS_INLINE_FUNCTION
 #if defined(__cpp_conditional_explicit) && \
@@ -1712,7 +681,7 @@ class View : public ViewTraits<DataType, Properties...> {
 #endif
           View(const typename Impl::MDSpanViewTraits<traits>::mdspan_type& mds,
                std::enable_if_t<
-                   !std::is_same_v<Impl::UnsupportedKokkosArrayLayout, U>>* =
+                   !std::is_same_v<Impl::UnsupportedKokkosArrayLayout, U> >* =
                    nullptr)
       : View(mds.data_handle(),
              Impl::array_layout_from_mapping<
@@ -1736,41 +705,6 @@ class View : public ViewTraits<DataType, Properties...> {
                                     AccessorType>& mds)
       : View(typename Impl::MDSpanViewTraits<traits>::mdspan_type(mds)) {
   }
-
-  //----------------------------------------
-  // Conversion to MDSpan
-  template <class OtherElementType, class OtherExtents, class OtherLayoutPolicy,
-            class OtherAccessor,
-            typename = std::enable_if_t<std::is_assignable_v<
-                mdspan<OtherElementType, OtherExtents, OtherLayoutPolicy,
-                       OtherAccessor>,
-                typename Impl::MDSpanViewTraits<traits>::mdspan_type>>>
-  KOKKOS_INLINE_FUNCTION constexpr operator mdspan<
-      OtherElementType, OtherExtents, OtherLayoutPolicy, OtherAccessor>() {
-    using mdspan_type = typename Impl::MDSpanViewTraits<traits>::mdspan_type;
-    return mdspan_type{data(),
-                       Impl::mapping_from_view_mapping<mdspan_type>(m_map)};
-  }
-
-  template <class OtherAccessorType = Impl::SpaceAwareAccessor<
-                typename traits::memory_space,
-                Kokkos::default_accessor<typename traits::value_type>>,
-            typename = std::enable_if_t<std::is_assignable_v<
-                typename traits::value_type*&,
-                typename OtherAccessorType::data_handle_type>>>
-  KOKKOS_INLINE_FUNCTION constexpr auto to_mdspan(
-      const OtherAccessorType& other_accessor =
-          typename Impl::MDSpanViewTraits<traits>::accessor_type()) {
-    using mdspan_type = typename Impl::MDSpanViewTraits<traits>::mdspan_type;
-    using ret_mdspan_type =
-        mdspan<typename mdspan_type::element_type,
-               typename mdspan_type::extents_type,
-               typename mdspan_type::layout_type, OtherAccessorType>;
-    return ret_mdspan_type{data(),
-                           Impl::mapping_from_view_mapping<mdspan_type>(m_map),
-                           other_accessor};
-  }
-#endif  // KOKKOS_ENABLE_IMPL_MDSPAN
 };
 
 template <typename D, class... P>
@@ -1794,7 +728,7 @@ template <unsigned N, typename... Args>
 KOKKOS_FUNCTION std::enable_if_t<
     N == View<Args...>::rank() &&
         std::is_same<typename ViewTraits<Args...>::specialize, void>::value,
-    View<Args...>>
+    View<Args...> >
 as_view_of_rank_n(View<Args...> v) {
   return v;
 }
@@ -1806,7 +740,7 @@ KOKKOS_FUNCTION std::enable_if_t<
     N != View<T, Args...>::rank() &&
         std::is_same<typename ViewTraits<T, Args...>::specialize, void>::value,
     View<typename RankDataType<typename View<T, Args...>::value_type, N>::type,
-         Args...>>
+         Args...> >
 as_view_of_rank_n(View<T, Args...>) {
   Kokkos::abort("Trying to get at a View of the wrong rank");
   return {};
@@ -1820,36 +754,6 @@ void apply_to_view_of_static_rank(Function&& f, View<Args...> a) {
 }  // namespace Impl
 //----------------------------------------------------------------------------
 //----------------------------------------------------------------------------
-
-namespace Impl {
-template <class ValueType, class TypeList>
-struct TypeListToViewTraits;
-
-template <class ValueType, class... Properties>
-struct TypeListToViewTraits<ValueType, Kokkos::Impl::type_list<Properties...>> {
-  using type = ViewTraits<ValueType, Properties...>;
-};
-
-// It is not safe to assume that subviews of views with the Aligned memory trait
-// are also aligned. Hence, just remove that attribute for subviews.
-template <class D, class... P>
-struct RemoveAlignedMemoryTrait {
- private:
-  using type_list_in  = Kokkos::Impl::type_list<P...>;
-  using memory_traits = typename ViewTraits<D, P...>::memory_traits;
-  using type_list_in_wo_memory_traits =
-      typename Kokkos::Impl::type_list_remove_first<memory_traits,
-                                                    type_list_in>::type;
-  using new_memory_traits =
-      Kokkos::MemoryTraits<memory_traits::impl_value & ~Kokkos::Aligned>;
-  using new_type_list = typename Kokkos::Impl::concat_type_list<
-      type_list_in_wo_memory_traits,
-      Kokkos::Impl::type_list<new_memory_traits>>::type;
-
- public:
-  using type = typename TypeListToViewTraits<D, new_type_list>::type;
-};
-}  // namespace Impl
 
 template <class D, class... P, class... Args>
 KOKKOS_INLINE_FUNCTION auto subview(const View<D, P...>& src, Args... args) {
@@ -1918,113 +822,6 @@ KOKKOS_INLINE_FUNCTION bool operator!=(const View<LT, LP...>& lhs,
 
 } /* namespace Kokkos */
 
-//----------------------------------------------------------------------------
-//----------------------------------------------------------------------------
-
-namespace Kokkos {
-namespace Impl {
-
-template <class Specialize, typename A, typename B>
-struct CommonViewValueType;
-
-template <typename A, typename B>
-struct CommonViewValueType<void, A, B> {
-  using value_type = std::common_type_t<A, B>;
-};
-
-template <class Specialize, class ValueType>
-struct CommonViewAllocProp;
-
-template <class ValueType>
-struct CommonViewAllocProp<void, ValueType> {
-  using value_type        = ValueType;
-  using scalar_array_type = ValueType;
-
-  template <class... Views>
-  KOKKOS_INLINE_FUNCTION CommonViewAllocProp(const Views&...) {}
-};
-
-template <class... Views>
-struct DeduceCommonViewAllocProp;
-
-// Base case must provide types for:
-// 1. specialize  2. value_type  3. is_view  4. prop_type
-template <class FirstView>
-struct DeduceCommonViewAllocProp<FirstView> {
-  using specialize = typename FirstView::traits::specialize;
-
-  using value_type = typename FirstView::traits::value_type;
-
-  enum : bool { is_view = is_view<FirstView>::value };
-
-  using prop_type = CommonViewAllocProp<specialize, value_type>;
-};
-
-template <class FirstView, class... NextViews>
-struct DeduceCommonViewAllocProp<FirstView, NextViews...> {
-  using NextTraits = DeduceCommonViewAllocProp<NextViews...>;
-
-  using first_specialize = typename FirstView::traits::specialize;
-  using first_value_type = typename FirstView::traits::value_type;
-
-  enum : bool { first_is_view = is_view<FirstView>::value };
-
-  using next_specialize = typename NextTraits::specialize;
-  using next_value_type = typename NextTraits::value_type;
-
-  enum : bool { next_is_view = NextTraits::is_view };
-
-  // common types
-
-  // determine specialize type
-  // if first and next specialize differ, but are not the same specialize, error
-  // out
-  static_assert(!(!std::is_same<first_specialize, next_specialize>::value &&
-                  !std::is_void<first_specialize>::value &&
-                  !std::is_void<next_specialize>::value),
-                "Kokkos DeduceCommonViewAllocProp ERROR: Only one non-void "
-                "specialize trait allowed");
-
-  // otherwise choose non-void specialize if either/both are non-void
-  using specialize = std::conditional_t<
-      std::is_same<first_specialize, next_specialize>::value, first_specialize,
-      std::conditional_t<(std::is_void<first_specialize>::value &&
-                          !std::is_void<next_specialize>::value),
-                         next_specialize, first_specialize>>;
-
-  using value_type = typename CommonViewValueType<specialize, first_value_type,
-                                                  next_value_type>::value_type;
-
-  enum : bool { is_view = (first_is_view && next_is_view) };
-
-  using prop_type = CommonViewAllocProp<specialize, value_type>;
-};
-
-}  // end namespace Impl
-
-template <class... Views>
-using DeducedCommonPropsType =
-    typename Impl::DeduceCommonViewAllocProp<Views...>::prop_type;
-
-// This function is required in certain scenarios where users customize
-// Kokkos View internals. One example are dynamic length embedded ensemble
-// types. The function is used to propagate necessary information
-// (like the ensemble size) when creating new views.
-// However, most of the time it is called with a single view.
-// Furthermore, the propagated information is not just for view allocations.
-// From what I can tell, the type of functionality provided by
-// common_view_alloc_prop is the equivalent of propagating accessors in mdspan,
-// a mechanism we will eventually use to replace this clunky approach here, when
-// we are finally mdspan based.
-// TODO: get rid of this when we have mdspan
-template <class... Views>
-KOKKOS_INLINE_FUNCTION DeducedCommonPropsType<Views...> common_view_alloc_prop(
-    Views const&... views) {
-  return DeducedCommonPropsType<Views...>(views...);
-}
-
-}  // namespace Kokkos
-
 #include <impl/Kokkos_ViewUniformType.hpp>
 #include <impl/Kokkos_Atomic_View.hpp>
 
@@ -2032,3 +829,4 @@ KOKKOS_INLINE_FUNCTION DeducedCommonPropsType<Views...> common_view_alloc_prop(
 //----------------------------------------------------------------------------
 
 #endif /* #ifndef KOKKOS_VIEW_HPP */
+#endif /* KOKKOS_ENABLE_IMPL_VIEW_LEGACY */
