+++ core/unit_test/TestOther.hpp
@@ -19,5 +19,7 @@
 #include <TestMemoryPool.hpp>
 #include <TestCXX11.hpp>
 
+#ifdef KOKKOS_ENABLE_IMPL_VIEW_LEGACY
 #include <TestViewCtorPropEmbeddedDim.hpp>
+#endif
 #endif
